requirements.txt
Django==5.1.2
djangorestframework==3.15.2
drf-yasg==1.21.7
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.4.0
psycopg2-binary==2.9.9
django-filter==24.3